Changed context to be in componets directory
